feat: move domain objects into shared package
