+++ backend/src/main/scala/model/db/BitBucketUserDB.scala
@@ -1,6 +1,8 @@
 package org.treemage
 package model.db
 
+import shared.model.domain.BitBucketUser
+
 import zio.schema.{DeriveSchema, Schema}
 
 import java.util.UUID
@@ -13,3 +15,18 @@ case class BitBucketUserDB(
 
 object BitBucketUserDB:
   given Schema[BitBucketUserDB] = DeriveSchema.gen[BitBucketUserDB]
+
+  def fromDomain(user: BitBucketUser): BitBucketUserDB =
+    BitBucketUserDB(
+      user.id,
+      user.name,
+      user.accountId
+    )
+
+  extension (self: BitBucketUserDB)
+    def toDomain: BitBucketUser =
+      BitBucketUser(
+        self.id,
+        self.name,
+        self.accountId
+      )
