+++ backend/src/main/scala/model/response/bitbucket/BitBucketUserResponse.scala
@@ -1,6 +1,8 @@
 package org.treemage
 package model.response.bitbucket
 
+import shared.model.domain.BitBucketUser
+
 import zio.schema.{DeriveSchema, Schema}
 
 case class BitBucketUserResponse(
@@ -11,3 +13,12 @@ case class BitBucketUserResponse(
 
 object BitBucketUserResponse:
   given Schema[BitBucketUserResponse] = DeriveSchema.gen[BitBucketUserResponse]
+
+  extension (self: BitBucketUserResponse)
+    def toDomain: Option[BitBucketUser] =
+      for id <- parseBitBucketUUID(self.uuid)
+      yield BitBucketUser(
+        id,
+        self.display_name,
+        self.account_id
+      )
