+++ backend/src/main/scala/service/CrawlStateServiceLive.scala
@@ -1,8 +1,8 @@
 package org.treemage
 package service
 
-import model.domain.crawl.{CrawlState, CrawlStatus}
 import repository.CrawlStateRepository
+import shared.model.domain.crawl.{CrawlState, CrawlStatus}
 
 import zio.*
