+++ backend/src/main/scala/service/UserService.scala
@@ -1,7 +1,7 @@
 package org.treemage
 package service
 
-import model.domain.BitBucketUser
+import shared.model.domain.BitBucketUser
 
 import zio.*
