+++ backend/src/main/scala/service/UserServiceLive.scala
@@ -1,8 +1,9 @@
 package org.treemage
 package service
 
-import model.domain.BitBucketUser
+import model.db.BitBucketUserDB
 import repository.BitBucketUserRepository
+import shared.model.domain.BitBucketUser
 
 import zio.*
 
@@ -12,13 +13,13 @@ case class UserServiceLive(userRepository: BitBucketUserRepository)
     extends UserService:
   override def getById(id: UUID): ZIO[Any, Nothing, Option[BitBucketUser]] =
     for user <- userRepository.getById(id).orDie
-    yield user.map(BitBucketUser.fromDB)
+    yield user.map(_.toDomain)
 
   override def getAll: ZIO[Any, Nothing, List[BitBucketUser]] =
-    userRepository.getAll.orDie.map(_.map(BitBucketUser.fromDB))
+    userRepository.getAll.orDie.map(_.map(_.toDomain))
 
   override def createOrUpdate(user: BitBucketUser): ZIO[Any, Nothing, UUID] =
-    userRepository.createOrUpdate(user.toDB).orDie
+    userRepository.createOrUpdate(BitBucketUserDB.fromDomain(user)).orDie
 
 object UserServiceLive:
   val layer: URLayer[BitBucketUserRepository, UserServiceLive] =
