+++ shared/src/main/scala/shared/model/domain/BitBucketUser.scala
@@ -1,0 +1,6 @@
+package org.treemage
+package shared.model.domain
+
+import java.util.UUID
+
+case class BitBucketUser(id: UUID, name: String, accountId: String)
