+++ shared/src/main/scala/shared/model/domain/pullrequest/PullRequest.scala
@@ -1,0 +1,16 @@
+package org.treemage
+package shared.model.domain.pullrequest
+
+import shared.model.domain.BitBucketUser
+import shared.model.domain.pullrequest.PullRequestState
+
+import java.time.LocalDateTime
+
+case class PullRequest(
+    id: Int,
+    title: String,
+    state: PullRequestState,
+    createdAt: LocalDateTime,
+    updatedAt: LocalDateTime,
+    author: BitBucketUser
+)
